Cast "threshold" from Int type to long type

+++ library/src/main/java/hotchemi/android/rate/AppRate.java
@@ -54,7 +54,7 @@ public final class AppRate {
     }
 
     private static boolean isOverDate(long targetDate, int threshold) {
-        return new Date().getTime() - targetDate >= threshold * 24 * 60 * 60 * 1000;
+        return new Date().getTime() - targetDate >= (long)threshold * 24 * 60 * 60 * 1000;
     }
 
     public AppRate setLaunchTimes(int launchTimes) {
